+++ packages/server/src/controllers/oauth2Provider.ts
@@ -1,16 +1,35 @@
-import { createHash, randomBytes } from 'crypto';
-import { URL } from 'url';
+import { createHash } from 'crypto';
 
-import { forbidden, notFound } from '@hapi/boom';
-import { addMinutes } from 'date-fns';
+import { badRequest, forbidden, notFound } from '@hapi/boom';
 import { Op } from 'sequelize';
 
-import { App, EmailAuthorization, OAuth2AuthorizationCode, User } from '../models';
+import { App, EmailAuthorization, Member, OAuth2Consent, User } from '../models';
 import type { KoaContext } from '../types';
+import { createOAuth2AuthorizationCode } from '../utils/model';
+import { hasScope } from '../utils/oauth2';
 
 interface Params {
   appId: number;
   redirectUri: string;
+}
+
+async function checkIsAllowed(app: App, user: User): Promise<boolean> {
+  const policy = app.definition?.security?.default?.policy ?? 'everyone';
+  if (policy === 'everyone') {
+    return true;
+  }
+
+  if (policy === 'invite' && !app.Users.length) {
+    return false;
+  }
+
+  if (policy === 'organization') {
+    return Boolean(
+      await Member.count({
+        where: { OrganizationId: app.OrganizationId, UserId: user.id },
+      }),
+    );
+  }
 }
 
 export async function getUserInfo(ctx: KoaContext<Params>): Promise<void> {
@@ -53,36 +72,71 @@ export async function getUserInfo(ctx: KoaContext<Params>): Promise<void> {
   };
 }
 
-export async function createAuthorizationCode(ctx: KoaContext<Params>): Promise<void> {
+export async function verifyOAuth2Consent(ctx: KoaContext<Params>): Promise<void> {
   const {
-    argv: { host },
+    argv,
     request: {
       body: { appId, redirectUri, scope },
     },
-    user: { id },
+    user,
   } = ctx;
 
-  const app = await App.findByPk(appId, { attributes: ['domain', 'path', 'OrganizationId'] });
+  const app = await App.findByPk(appId, {
+    attributes: ['definition', 'domain', 'id', 'path', 'OrganizationId'],
+    include: [
+      { model: OAuth2Consent, where: { UserId: user.id }, required: false },
+      { model: User, where: { id: user.id }, required: false },
+    ],
+  });
 
   if (!app) {
     throw notFound('App not found');
   }
 
-  const appHost = `${app.path}.${app.OrganizationId}.${new URL(host).hostname}`;
-  const redirectHost = new URL(redirectUri).hostname;
-  if (redirectHost !== appHost && redirectHost !== app.domain) {
-    throw forbidden('Invalid redirectUri');
+  const isAllowed = await checkIsAllowed(app, user);
+
+  if (!app.OAuth2Consents?.length || !hasScope(app.OAuth2Consents[0].scope, scope)) {
+    throw badRequest('User has not agreed to the requested scopes', {
+      appName: app.definition.name,
+      isAllowed,
+    });
   }
 
-  const { code } = await OAuth2AuthorizationCode.create({
-    AppId: appId,
-    code: randomBytes(12).toString('hex'),
-    expires: addMinutes(new Date(), 10),
-    redirectUri,
-    scope,
-    UserId: id,
+  if (!isAllowed) {
+    throw badRequest('User is not allowed to login due to the app’s security policy', {
+      appName: app.definition.name,
+      isAllowed,
+    });
+  }
+
+  ctx.body = await createOAuth2AuthorizationCode(argv, app, redirectUri, scope, user);
+}
+
+export async function agreeOAuth2Consent(ctx: KoaContext<Params>): Promise<void> {
+  const {
+    argv,
+    request: {
+      body: { appId, redirectUri, scope },
+    },
+    user,
+  } = ctx;
+
+  const app = await App.findByPk(appId, {
+    attributes: ['domain', 'definition', 'id', 'path', 'OrganizationId'],
+    include: [OAuth2Consent, { model: User, where: { id: user.id }, required: false }],
   });
-  ctx.body = {
-    code,
-  };
+
+  if (!app) {
+    throw notFound('App not found');
+  }
+
+  if (!(await checkIsAllowed(app, user))) {
+    throw badRequest('User is not allowed to login due to the app’s security policy', {
+      appName: app.definition.name,
+      isAllowed: false,
+    });
+  }
+
+  await OAuth2Consent.upsert({ AppId: appId, UserId: user.id, scope });
+  ctx.body = await createOAuth2AuthorizationCode(argv, app, redirectUri, scope, user);
 }