+++ packages/utils/src/api/components/responses/oauth2AuthorizationCode.ts
@@ -1,0 +1,12 @@
+import type { OpenAPIV3 } from 'openapi-types';
+
+export const oauth2AuthorizationCode: OpenAPIV3.ResponseObject = {
+  description: 'An OAuth2 authorization code.',
+  content: {
+    'application/json': {
+      schema: {
+        $ref: '#/components/schemas/App',
+      },
+    },
+  },
+};